sl-rexi scripts

+++ benchmarks_plane/sl-rexi/unstablejet_dif/jobs_run.sh
@@ -17,11 +17,14 @@ fi
 
 
 for i in $DIRS; do
-	test -d "$i" || continue
-
+    test -d "$i" || continue
+    j="$i"".out"
+    if [ ! -f "$j" ]; then
 	cd "$i"
+	echo $i
 	./run.sh | tee "../$i.out"
 	#test ${PIPESTATUS[0]} -eq 0 || exit 2>&1
 	cd ".."
+    fi
 
 done
